pyproject.toml
[project]
name = "rbc-fno-surrogate"
version = "1.0.0"
description = "FNO-Based Surrogate model for Rayleigh-Benard Convection"
readme = "README.md"
authors = [
    { email = "tmarkmann@techfak.uni-bielefeld.de" }
]
requires-python = ">=3.11"
dependencies = [
    "ffmpeg>=1.4",
    "h5py>=3.14.0",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "lightning[extra]==2.5.5",
    "neuraloperator==1.0.2",
    "pandas>=2.3.0",
    "pyqt5>=5.15.11",
    "rbc-gym",
    "rich>=14.0.0",
    "seaborn>=0.13.2",
    "torch==2.7.1",
    "torch-harmonics==0.7.3",
    "torchmetrics==1.8.2",
    "torchvision==0.22.1",
    "wandb==0.20.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.8",
    "ruff>=0.11.6",
]

[tool.uv.sources]
rbc-gym = { git = "https://github.com/HammerLabML/RBC-Gym" }

[tool.hatch.build.targets.wheel]
packages = ["src/rbc_fno_surrogate"]
